Split view model to separate file
